Added pagination, sorting, and searching using Backgrid
Details:
Moved the web client stuff into a 'static' folder. Added a higher
verbosity level for heartbeat statements. Since live reload didn't
mesh well with pagination (it's weird to see elements moving off the
page when you're just downloading stuff), I swapped that for a refresh
button, which will also return to the first page. I had to jam this
refresh button into the backgrid filter, since resizing issues made
putting it beside the filter hard (explained in
backbone-filter-custom.js). There was a bug with the backbone-pageable
when combined with the filter, where it would throw an error when your
search term narrowed the number of pages past your current page. This
was fixed in the newest version of backbone-pageable, but that made
the backgrid-paginator stop updating it's page handlers when you
clicked a new page, so I hacked the paginator to fix that (explained
in backgrid-paginator-custom.js). I also replaced the minified
javascript libraries with documented ones, for debugging convenience.