Added downloading directories using tarfiles.
It creates a tar in memory every time a directory is queried for. I
modified the heartbeat to add directories. I also removed
dotfiles/dotdirectories from the movie listing, because they are
generally unneeded and create clutter. Also I refactored the server to
use filepath.Join in more places, and I beefed up testing for the movie
table handler and added tests for directories.